Only run the runLogin function once, conditional on an email and password existing in local storage

+++ pages/new-employer-signup/index.js
@@ -4,10 +4,11 @@ import Head from "next/head"
 const NewEmployerSignup = () => {
 
   const runLogin = () => {
-    const t = {'action':'login','username': localStorage.getItem("email"), 'password': localStorage.getItem("password")};
-    document.getElementById('rb').onload = null;
-    document.getElementById('rb').contentWindow.postMessage(t, 'https://www.redballoon.work/login/employer/purchase');
-    localStorage.clear()
+    if(localStorage.getItem("email") && localStorage.getItem("password")){
+      const t = {'action':'login','username': localStorage.getItem("email"), 'password': localStorage.getItem("password")};
+      document.getElementById('rb').contentWindow.postMessage(t, 'https://www.redballoon.work/login/employer/purchase');
+      localStorage.clear()
+    }
   }
 
   return (
